+++ src/modules/controls/overviewMap/store/indexOverviewMap.js
@@ -1,0 +1,10 @@
+import getters from "./gettersOverviewMap";
+import mutations from "./mutationsOverviewMap";
+import state from "./stateOverviewMap";
+
+export default {
+    namespaced: true,
+    state,
+    mutations,
+    getters
+};