+++ src/modules/controls/overviewMap/store/stateOverviewMap.js
@@ -1,0 +1,6 @@
+const state = {
+    overviewLayer: undefined
+};
+
+export default state;
+